Finalizing messages apis

+++ app/Http/Controllers/API/VideosController.php
@@ -2,7 +2,6 @@
 
 namespace App\Http\Controllers\API;
 
-use App\News;
 use App\Video;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
@@ -52,10 +51,10 @@ class VideosController extends Controller
         $t = $request->all();
         $t->verified = false;
         $t->user_id = Auth::user()->id;
-        $news = new News($t);
+        $news = new Video($t);
         $data['statues'] = "200 Ok";
         $data['error'] = null;
-        $data['data']['videos'] = $news;
+        $data['data']['video'] = $news;
         return response()->json($data, 200);
     }
 
@@ -68,17 +67,28 @@ class VideosController extends Controller
     public function show($id)
     {
         if (Bouncer::is(Auth::user())->an('admin', 'admin_level_1', 'admin_level_2')) {
-            $news = Video::findOrFail($id);
+            $news = Video::find($id);
             $data['statues'] = "200 Ok";
             $data['error'] = null;
-            $data['data']['videos'] = $news;
-            return response()->json($data, 200);
+            $data['data']['video'] = $news;
+            if ($news == null) {
+                return response()->json($data, 404);
+            } else {
+                return response()->json($data, 200);
+            }
         } else {
             $news = DB::table('videos')->where('verified', true)->where('id', $id)->get()->first();
             $data['statues'] = "200 Ok";
             $data['error'] = null;
-            $data['data']['videos'] = $news;
-            return response()->json($data, 200);
+            $data['data']['video'] = $news;
+            if ($news == null) {
+                $data['statues'] = "404 not found";
+                $data['error'] = "Not found";
+                $data['data'] = null;
+                return response()->json($data, 404);
+            } else {
+                return response()->json($data, 200);
+            }
         }
 
     }
@@ -94,7 +104,7 @@ class VideosController extends Controller
     public function update(Request $request, $id)
     {
         $t = $request->all();
-        $news = Video::findOrFail($id);
+        $news = Video::find($id);
         if ($news->user_id != Auth::user_id && Bouncer::is(Auth::user())->notAn('admin')) {
             $data['statues'] = "401 unauthorized";
             $data['error'] = "UnAuthorized";
@@ -102,11 +112,19 @@ class VideosController extends Controller
             return response()->json($data, 401);
         }
         $t->user_id = Auth::user()->id;
-        $news->update($request);
-        $data['statues'] = "200 Ok";
-        $data['error'] = null;
-        $data['data']['videos'] = $news;
-        return response()->json($data, 200);
+        if ($news == null) {
+            $data['statues'] = "404 not found";
+            $data['error'] = "Not found";
+            $data['data'] = null;
+            return response()->json($data, 404);
+        } else {
+            $news->update($t);
+            $data['statues'] = "200 Ok";
+            $data['error'] = null;
+            $data['data']['video'] = $news;
+            return response()->json($data, 200);
+        }
+
     }
 
     /**
@@ -117,18 +135,26 @@ class VideosController extends Controller
      */
     public function destroy($id)
     {
-        $news = Video::findOrFail($id);
+        $news = Video::find($id);
         if ($news->user_id != Auth::user_id && Bouncer::is(Auth::user())->notAn('admin')) {
             $data['statues'] = "401 unauthorized";
             $data['error'] = "UnAuthorized";
             $data['data'] = null;
             return response()->json($data, 401);
         }
-        $news->destroy();
-        $data['statues'] = "200 Ok";
-        $data['error'] = null;
-        $data['data'] = null;
-        return response()->json($data, 200);
+        if ($news == null) {
+            $data['statues'] = "404 not found";
+            $data['error'] = "Not found";
+            $data['data'] = null;
+            return response()->json($data, 404);
+        } else {
+            $news->destroy();
+            $data['statues'] = "200 Ok";
+            $data['error'] = null;
+            $data['data'] = null;
+            return response()->json($data, 200);
+        }
+
     }
 
     /**
@@ -139,11 +165,19 @@ class VideosController extends Controller
      */
     public function verify($id)
     {
-        $news = Video::findOrFail($id);
-        $news->verified = true;
-        $data['statues'] = "200 Ok";
-        $data['error'] = null;
-        $data['data'] = $news;
-        return response()->json($data, 200);
+        $news = Video::find($id);
+        if ($news == null) {
+            $data['statues'] = "404 not found";
+            $data['error'] = "Not found";
+            $data['data'] = null;
+            return response()->json($data, 404);
+        } else {
+            $news->verified = true;
+            $data['statues'] = "200 Ok";
+            $data['error'] = null;
+            $data['data'] = $news;
+            return response()->json($data, 200);
+        }
+
     }
 }
